+++ packages/terra-tabs/CHANGELOG.md
@@ -2,8 +2,18 @@
 
 ## Unreleased
 
+## 7.8.0 - (September 21, 2023)
+
+* Changed
+  * Updated code for tab active indicator issue.
+
+## 7.7.0 - (September 19, 2023)
+
 * Added
   * Added prop to support the focus traverse within the tab content.
+
+* Fixed
+  * Fixed tab active indicator issue.
 
 ## 7.6.0 - (September 5, 2023)